requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiostream==0.4.5
anyio==3.6.1
astor==0.8.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
colorlog==6.7.0
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
datasets==2.4.0
decorator==5.1.1
dill==0.3.4
docarray==0.16.2
docker==6.0.0
fastapi==0.81.0
faster-tokenizer==0.1.6
filelock==3.8.0
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-reflection==1.47.0
h11==0.13.0
httptools==0.4.0
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
jcloud==0.0.35
jieba==0.42.1
jina==3.8.1
jina-hubble-sdk==0.15.4
joblib==1.1.0
kiwisolver==1.4.4
lz4==4.0.2
matplotlib==3.5.3
multidict==6.0.2
multiprocess==0.70.12.2
numpy==1.23.2
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddle2onnx==0.9.7
paddlefsl==1.1.0
paddlenlp==2.3.7
paddlepaddle==2.3.2
pandas==1.4.4
pathspec==0.10.0
Pillow==9.2.0
prometheus-client==0.14.1
protobuf==3.20.0
pyarrow==9.0.0
pycparser==2.21
pydantic==1.10.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
responses==0.18.0
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.1
sentencepiece==0.1.97
seqeval==1.2.2
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websocket-client==1.4.0
websockets==10.3
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
